trilha1/curso8/requirements.txt
fastmcp>=2.8.0
wikipedia
openai
python-dotenv
requests